+++ src/main/java/com/twilio/rest/api/v2010/account/conference/ParticipantFetcher.java
@@ -27,7 +27,8 @@ public class ParticipantFetcher extends Fetcher<Participant> {
      *
      * @param pathConferenceSid The SID of the conference with the participant to
      *                          fetch
-     * @param pathCallSid The Call SID of the resource to fetch
+     * @param pathCallSid The Call SID or URL encoded label of the participant to
+     *                    fetch
      */
     public ParticipantFetcher(final String pathConferenceSid,
                               final String pathCallSid) {
@@ -42,7 +43,8 @@ public class ParticipantFetcher extends Fetcher<Participant> {
      *                       fetch
      * @param pathConferenceSid The SID of the conference with the participant to
      *                          fetch
-     * @param pathCallSid The Call SID of the resource to fetch
+     * @param pathCallSid The Call SID or URL encoded label of the participant to
+     *                    fetch
      */
     public ParticipantFetcher(final String pathAccountSid,
                               final String pathConferenceSid,