+++ src/main/java/com/twilio/rest/notify/v1/ServiceCreator.java
@@ -34,6 +34,8 @@ public class ServiceCreator extends Creator<Service> {
     private Boolean logEnabled;
     private String alexaSkillId;
     private String defaultAlexaNotificationProtocolVersion;
+    private String deliveryCallbackUrl;
+    private Boolean deliveryCallbackEnabled;
 
     /**
      * A descriptive string that you create to describe the resource. It can be up
@@ -189,6 +191,28 @@ public class ServiceCreator extends Creator<Service> {
     }
 
     /**
+     * URL to send delivery status callback..
+     *
+     * @param deliveryCallbackUrl Webhook URL
+     * @return this
+     */
+    public ServiceCreator setDeliveryCallbackUrl(final String deliveryCallbackUrl) {
+        this.deliveryCallbackUrl = deliveryCallbackUrl;
+        return this;
+    }
+
+    /**
+     * Callback configuration that enables delivery callbacks, default false.
+     *
+     * @param deliveryCallbackEnabled Enable delivery callbacks
+     * @return this
+     */
+    public ServiceCreator setDeliveryCallbackEnabled(final Boolean deliveryCallbackEnabled) {
+        this.deliveryCallbackEnabled = deliveryCallbackEnabled;
+        return this;
+    }
+
+    /**
      * Make the request to the Twilio API to perform the create.
      *
      * @param client TwilioRestClient with which to make the request
@@ -271,6 +295,14 @@ public class ServiceCreator extends Creator<Service> {
 
         if (defaultAlexaNotificationProtocolVersion != null) {
             request.addPostParam("DefaultAlexaNotificationProtocolVersion", defaultAlexaNotificationProtocolVersion);
+        }
+
+        if (deliveryCallbackUrl != null) {
+            request.addPostParam("DeliveryCallbackUrl", deliveryCallbackUrl);
+        }
+
+        if (deliveryCallbackEnabled != null) {
+            request.addPostParam("DeliveryCallbackEnabled", deliveryCallbackEnabled.toString());
         }
     }
 }