+++ src/main/java/com/twilio/rest/supersim/v1/FleetUpdater.java
@@ -51,7 +51,7 @@ public class FleetUpdater extends Updater<Fleet> {
 
     /**
      * The SID or unique name of the Network Access Profile that will control which
-     * cellular network operators the Fleet's SIMs can connect to.
+     * cellular networks the Fleet's SIMs can connect to.
      *
      * @param networkAccessProfile The SID or unique name of the Network Access
      *                             Profile of the Fleet