+++ CHANGES.md
@@ -1,6 +1,49 @@
 twilio-java changelog
 =====================
 
+[2020-06-10] Version 7.52.0
+---------------------------
+**Library - Docs**
+- [PR #533](https://github.com/twilio/twilio-java/pull/533): link to custom HTTP client instructions. Thanks to [@thinkingserious](https://github.com/thinkingserious)!
+- [PR #532](https://github.com/twilio/twilio-java/pull/532): link to descriptive exception types. Thanks to [@thinkingserious](https://github.com/thinkingserious)!
+
+**Library - Chore**
+- [PR #531](https://github.com/twilio/twilio-java/pull/531): drop the region being passed to each request in the REST object. Thanks to [@childish-sambino](https://github.com/childish-sambino)!
+
+**Library - Feature**
+- [PR #530](https://github.com/twilio/twilio-java/pull/530): add regional and edge support. Thanks to [@childish-sambino](https://github.com/childish-sambino)!
+
+**Api**
+- Added `pstnconnectivity` to `usage_record` API
+
+**Autopilot**
+- Add dialogue_sid param to Query list resource
+
+**Contacts**
+- Added contacts channel create, update, read, delete API definition
+- Added contacts Get API definition
+- Added contacts Update API definition
+- Added contacts Delete API definition
+- Added contacts Search API definition
+- Added AccountSID to CFD CREATE and GET Responses
+- Added contacts location Create, update, read and delete API definitions
+
+**Notify**
+- delivery_callback_url and delivery_callback_enabled added
+
+**Numbers**
+- Add `provisionally-approved` as a Bundle status
+
+**Preview**
+- `BrandsInformation` endpoint now returns a single `BrandsInformation`
+- Deleted phone number required field in the brand phone number endpoint from `kyc-api`
+- Removed insights `preview API` from API Definitions **(breaking change)**
+- Added `BrandsInformation` endpoint to query brands information stored in KYC
+
+**Supersim**
+- Require a Network Access Profile when creating a Fleet **(breaking change)**
+
+
 [2020-05-27] Version 7.51.0
 ---------------------------
 **Library - Fix**